packages/flask-py/requirements.txt
fastapi
uvicorn
langchain
langchain-openai
langchain-chroma
openai
pylint
langflow
python-dotenv
